requirements.txt
prometheus-client>=0.21
python-dotenv>=1.0
aiohttp>=3.10
asyncio>=3.4
pysmartthings==0.7.8